move utilities back to card renderer
